rename storage plugin file
