+++ frontend/src/store/index.js
@@ -69,7 +69,7 @@ import pick from 'lodash/pick'
 import moment from '@/utils/moment'
 import createSocketPlugin from './plugins/socketPlugin'
 import createMediaPlugin from './plugins/mediaPlugin'
-import createStoragePlugin from './plugins/storage'
+import createStoragePlugin from './plugins/storagePlugin'
 import shoots from './modules/shoots'
 import cloudProfiles from './modules/cloudProfiles'
 import gardenerExtensions from './modules/gardenerExtensions'
